% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfsl.R
\name{bfsl.control}
\alias{bfsl.control}
\title{Controls the iterations in bfsl.}
\usage{
bfsl.control(tol = 1e-10, maxiter = 100)
}
\arguments{
\item{tol}{A positive numeric value specifying the tolerance level for the
convergence criterion}

\item{maxiter}{A positive integer specifying the maximum number of iterations allowed.}
}
\value{
A \code{list} with two components:
\item{tol}{}
\item{maxiter}{}
}
\description{
\code{bfsl.control} allows the user to set some characteristics of the \code{bfsl}
best-fit straight line algorithm.
}
\examples{
bfsl.control(tol = 1e-8, maxiter = 1000)

}
\seealso{
\code{\link{bfsl}}
}
