% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfsl.R
\name{plot.bfsl}
\alias{plot.bfsl}
\title{Plot method for bfsl results}
\usage{
\method{plot}{bfsl}(x, grid = TRUE, ...)
}
\arguments{
\item{x}{An object of class "\code{bfsl}".}

\item{grid}{If \code{TRUE} (default) grid lines are plotted.}

\item{...}{Further parameters to be passed to the plotting routines.}
}
\description{
Plot method for objects of class "\code{bfsl}".
}
\details{
\code{plot.bfsl} plots the data points with error bars and the calculated
best-fit straight line.
}
