% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfsl.R
\name{predict.bfsl}
\alias{predict.bfsl}
\title{Predict Method for bfsl Model Fits}
\usage{
\method{predict}{bfsl}(
  x,
  newdata,
  interval = c("none", "confidence"),
  level = 0.95,
  se.fit = FALSE
)
}
\arguments{
\item{x}{Object of class \code{"bfsl"}.}

\item{newdata}{A data frame with variable \code{x} to predict.
If omitted, the fitted values are used.}

\item{interval}{Type of interval calculation. \code{"none"} or \code{"confidence"}.}

\item{level}{Confidence level.}

\item{se.fit}{A switch indicating if standard errors are returned.}
}
\value{
\code{predict.bfsl} produces a vector of predictions or a matrix of
predictions and bounds with column names \code{fit}, \code{lwr}, and \code{upr}
if interval is set to \code{"confidence"}.

If \code{se.fit} is \code{TRUE}, a list with the following components is returned:
\tabular{ll}{
\code{fit} \tab Vector or matrix as above \cr
\code{se.fit} \tab Standard error of predicted means
}
}
\description{
\code{predict.bfsl} predicts future values based on the bfsl fit.
}
