% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfsl.R
\name{print.bfsl}
\alias{print.bfsl}
\title{Print method for bfsl results}
\usage{
\method{print}{bfsl}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class "\code{bfsl}".}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Further arguments passed to \code{print.default}.}
}
\description{
\code{print} method for class "\code{bfsl}".
}
