% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfsl.R
\name{summary.bfsl}
\alias{summary.bfsl}
\title{Summary Method for bfsl Results}
\usage{
\method{summary}{bfsl}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{An object of class "\code{bfsl}".}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Further arguments passed to \code{summary.default}.}
}
\description{
\code{summary} method for class \code{"bfsl"}.
}
