% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfsl.R
\name{bfsl}
\alias{bfsl}
\title{Calculates the best-fit straight line.}
\usage{
bfsl(x, y, sd_x = 0, sd_y = 1, r = 0, control = bfsl.control())
}
\arguments{
\item{x}{A vector of \emph{x} observations.}

\item{y}{A vector of \emph{y} observations.}

\item{sd_x}{An optional vector or scalar of \emph{x} measurement error
standard deviation(s).}

\item{sd_y}{An optional vector or scalar of \emph{y} measurement error
standard deviation(s).}

\item{r}{An optional vector or scalar of correlation coefficient(s) between
errors in \emph{x} and \emph{y}.}

\item{control}{An optional list of control settings. See \code{\link{bfsl.control}}
for the names of the settable control values and their effect.}
}
\value{
A \code{list} with the components:
\item{coefficients}{A \code{2x2} matrix with columns of the fitted coefficients
(intercept and slope) and their standard errors.}
\item{chisq}{The goodness of fit.}
\item{control}{The control \code{list} used, see the \code{control} argument.}
\item{convInfo}{A \code{list} with convergence information.}
\item{call}{The matched call.}
\item{data}{A list containing \code{x}, \code{y}, \code{sd_x}, \code{sd_y}
and \code{r}.}
}
\description{
\code{bfsl} calculates the best-fit straight line to independent points with
(possibly correlated) normally distributed errors in both coordinates.
}
\details{
\code{bfsl} provides the general least-squares estimation solution to the
problem of fitting a straight line to independent data with (possibly
correlated) normally distributed errors in \code{x} and \code{y}. Setting
\code{sd_x = sd_y} and \code{r = 0} leads to the orthogonal distance
regression (ODR) solution, also known as major axis regression. Setting
\code{sd_x = var(x)}, \code{sd_y = var(y)} and \code{r = 0} leads to the
geometric mean regression (GMR) solution, also known as reduced major axis
regression or standardised major axis regression.

With \code{sd_x = 0} the (weigthed) ordinary least squares (OLS) solution is
obtained. The calculated standard errors of the slope and intercept multiplied
with \code{sqrt(chisq)} gives the OLS standard errors.

The goodness of fit metric \code{chisq} is a weighted reduced chi-squared
statistic. It compares the deviations of the points from the fit line to the
assigned measurement error standard deviations. If \code{x} and \code{y} are
indeed related by a straight line, and if the assigned measurement errors
are correct (and normally distributed), then \code{chisq} will equal 1. A
\code{chisq > 1} indicates underfitting: the fit does not fully capture the
data or the measurement errors have been underestimated. A \code{chisq < 1}
indicates overfitting: either the model is improperly fitting noise, or the
measurement errors have been overestimated.
}
\examples{
bfsl(pearson$x, pearson$y, pearson$sd_x, pearson$sd_y)

fit = bfsl(pearson$x, pearson$y, pearson$sd_x, pearson$sd_y)
plot(fit)

}
\references{
York, D. (1969). Least squares fitting of a straight line with
correlated errors. \emph{Earth and Planetary Science Letters}, 5, 320–324,
https://doi.org/10.1016/S0012-821X(68)80059-7
}
